services/auth-service/requirements.txt
flask==3.0.3
flask-limiter==3.8.0
PyJWT==2.8.0
bcrypt==4.1.3
boto3==1.34.131
psycopg2-binary==2.9.9
prometheus-client==0.20.0
python-json-logger==2.0.7
gunicorn==22.0.0
